+++ src/views/lighttable.c
@@ -42,6 +42,10 @@
 #include "views/view.h"
 #include "views/view_api.h"
 
+#ifdef USE_LUA
+#include "lua/image.h"
+#endif
+
 #include <assert.h>
 #include <dirent.h>
 #include <errno.h>
@@ -223,6 +227,40 @@ static gboolean _preview_get_state(dt_view_t *self)
   return lib->preview_state;
 }
 
+#ifdef USE_LUA
+
+static int set_image_visible_cb(lua_State *L)
+{
+  dt_lua_image_t imgid = -1;
+  if(luaL_testudata(L, 1, "dt_lua_image_t"))
+  {
+    luaA_to(L, dt_lua_image_t, &imgid, 1);
+     dt_thumbtable_set_offset_image(dt_ui_thumbtable(darktable.gui->ui), imgid, TRUE);
+    return 0;
+  }
+  else
+  {
+    return luaL_error(L, "no image specified");
+  }
+}
+
+static gboolean is_image_visible_cb(lua_State *L)
+{
+  dt_lua_image_t imgid = -1;
+  if(luaL_testudata(L, 1, "dt_lua_image_t"))
+  {
+    luaA_to(L, dt_lua_image_t, &imgid, 1);
+    lua_pushboolean(L, dt_thumbtable_check_imgid_visibility(dt_ui_thumbtable(darktable.gui->ui), imgid));
+    return 1;
+  }
+  else
+  {
+    return luaL_error(L, "no image specified");
+  }
+}
+
+#endif
+
 void init(dt_view_t *self)
 {
   self->data = calloc(1, sizeof(dt_library_t));
@@ -235,6 +273,20 @@ void init(dt_view_t *self)
 
   // ensure the memory table is up to date
   dt_collection_memory_update();
+  
+#ifdef USE_LUA
+  lua_State *L = darktable.lua_state.state;
+  const int my_type = dt_lua_module_entry_get_type(L, "view", self->module_name);
+  lua_pushlightuserdata(L, self);
+  lua_pushcclosure(L, set_image_visible_cb, 1);
+  dt_lua_gtk_wrap(L);
+  lua_pushcclosure(L, dt_lua_type_member_common, 1);
+  dt_lua_type_register_const_type(L, my_type, "set_image_visible");
+  lua_pushcclosure(L, is_image_visible_cb, 1);
+  dt_lua_gtk_wrap(L);
+  lua_pushcclosure(L, dt_lua_type_member_common, 1);
+  dt_lua_type_register_const_type(L, my_type, "is_image_visible");
+#endif
 }